move Tracing from block to trace module
